Database stuff

+++ app/src/main/java/com/example/spapp/models/Position.java
@@ -17,7 +17,7 @@ public class Position
     public String description;
 
     @ColumnInfo(name = "day")
-    public int day;
+    public long day;
 
     public int getId() {
         return id;
